Fix compiler warnings.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@12055 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ server/spell_effect.c
@@ -2849,7 +2849,7 @@ int cast_consecrate(object *op, object *caster, object *spell) {
             } else {
                 /* If we got here, we are consecrating an altar */
                 object *new_altar;
-                int letter;
+                size_t letter;
                 archetype *altar_arch;
 
                 snprintf(buf, MAX_BUF, "altar_");
